pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "nanoporethon"
version = "0.0.0"
description ="Nanopore Python program for analyzing Nanopore files"
dependencies = [
    "numpy",
    "pytest",
    "pytest-cov",
    "matplotlib",
    "h5py",
    "PyYAML",
    "jsonschema",
]

[tool.pytest.ini_options]
addopts = "--cov=nanoporethon --cov-report=term-missing"

[tool.coverage.run]
omit = [
    "*/nanoporethon/data_navi_gui.py",
    "*/nanoporethon/event_classifier_gui.py"
]

[tool.coverage.report]
fail_under = 65
